完成了编码器部分，基本完成了转向，但转向由于四驱出现抖动，待解决

+++ project/modules/inc/MOTOR.h
@@ -26,6 +26,6 @@ typedef struct {
 
 void Motor_Init(Motor_Handle *motor);
 void Motor_SetDirection(Motor_Handle *motor, Motor_DirectionType dir);
-void Motor_SetSpeed(Motor_Handle *motor, uint32_t percent);
+void Motor_SetSpeed(Motor_Handle *motor, int16_t percent);
 
 #endif
